requirements.txt
Flask==3.0.0
yt-dlp==2024.11.18
opencv-python-headless==4.10.0.84
Pillow==11.0.0
werkzeug==3.0.1
gunicorn==21.2.0
numpy==1.26.4
